feat(plugin-crx): add picker context to page-action protocol
